python-requirements.txt
# Python dependencies for AI Orchestrator

# Core
streamlit>=1.31.0
python-dotenv>=1.0.0

# AI Providers
openai>=1.12.0
google-generativeai>=0.3.2
anthropic>=0.18.0

# Local AI
ollama>=0.1.8

# Multi-agent framework
crewai>=0.30.11
langchain>=0.1.0

# Utilities
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0